+++ src/services/dashboard/learnersDashboard.service.ts
@@ -1,7 +1,7 @@
 import { ApiResponse } from "@/types/api";
 import { LearnersDashboardResponse } from "@/types/dashboard/learnersDashboard";
 
-import { axiosClientSVC } from "../axiosClient";
+import { axiosClientSVC } from "../../shared/common/api/axiosClient";
 
 export const getLearnersDashboardAPI = async () => {
   const { data } =
