+++ src/services/exams/assignment-results.service.test.ts
@@ -1,4 +1,4 @@
-import { axiosClient } from "@/services/axiosClient";
+import { axiosClient } from "@/shared/common/api/axiosClient";
 
 import {
   upsertAssignmentResultsAPI,
