+++ src/services/exams/statistics.service.ts
@@ -1,4 +1,4 @@
-import { axiosClient } from "@/services/axiosClient";
+import { axiosClient } from "@/shared/common/api/axiosClient";
 import { ApiResponse } from "@/types/api";
 import type { ExamStatisticsApi } from "@/types/exams";
