+++ src/services/instructorPost.service.ts
@@ -1,4 +1,4 @@
-import { axiosClient } from "@/services/axiosClient";
+import { axiosClient } from "@/shared/common/api/axiosClient";
 import {
   GetInstructorPostsResponse,
   GetInstructorPostDetailResponse,
